Move sync to multi
